Moved parts of the `MainConfig` to a dedicated `WebConfig`
Web-related thigs will go there from now on
